Modified Linux instructions
Added dependencies for Arch Linux. They are confirmed to work with a
fresh install of Arch Linux. They may also work on Manjaro Linux
(Arch-Based but the package updates runs a few weeks behind Arch).
Instructions to refer to step 4 of the Windows setup are also included,
as it seems to be one of the most common questions.