Formatting, ESRI Shapefiles, libshape
Fixed a workspace setting for VS Code that hopefully fixes formatting
issues others were having. Added the libshape library to process ESRI
Shapefiles in Grail. This could be replaced by a homebrewed version later,
if necessary. We also have to determine how to include it in the build,
as there is an outdated version available in the MSYS repositories.
Lastly, bindings for libshape were added so we can read and store
shape data to be drawn with Grail.