Importing Shapelib
The files believed to be the core of Shapelib have been manually
imported and added to Grail's compilation process. There probably should
be a way of grabbing these files from the github repository
automatically, for future builds. The library is compiled into
libs/libshape.a and then linked into the rest of the project from there.

Shapelib is licensed as shown here:
https://github.com/OSGeo/shapelib/blob/master/web/license.html

If we wanted to remove the library, we could probably rewind the commit
history from here and be safe. Instead, we would just have to have
Shapelib as a dependency required for Grail to compile (like FreeType,
and GLFW).

Of less important note, the TestESRIPolygon demo now is working and is
printing all points in the shape. The runtime could be reduced a bit
though. While the runtime isn't bad for smaller files, it takes about
seven seconds to load in the county shapefile. This is still just under
five million points though, so perhaps these load times are to be
expected. Development was also started on a ShapefileDemo, which will
hopefully implement the same functionality as MapDemo but using ESRI
Shapefiles.