Minor formatting and error handling
Did some minor reformatting in the socket code and modified the error
handling. Previously it was attempting to always print WSAGetLastError()
on a send/recv failure, but this isn't compatible with the existing
Linux code. Instead some preprocessor trickery was used to use errno,but
redefine errno to WSAGetLastError() on Windows systems.