Fixed clang_format config
The previous config for clang_format had a copy of the json key in it,
as well as some escaped quotes. The malformed json value made it so the
formatter never actually ran. This should now be fixed. Optionally,
developers may want to turn enable the "Format on Save" setting.