Improve ESRI-based classes and add Doxygen config
ESRIPolygons are now using vectors of ESRIPoints. This may be changed
later to improve speed and compatibility with the drawing functions.

Modifies ESRI Shapefile demos to reflect the change to ESRIPolygons.

Removes the filtering from StyledMultiShape's drawPolygon, as it was
only being used for the one demo. Further investigation is needed into
the implementation of this function. In its current state, it tends
to draw lines between vertices that should not have lines.

Adds configuration for Doxygen. This is not officially supported in
Grail at the moment, but it currently generates documentation for the
project for docbook, latex, man, and html-based processing. The
.gitignore has been updated as a result, to prevent the generated
documentation from being pushed into the repository. This may be changed
upon further revision.