fixes FreeType on Windows, adds frequent issues
The README now has a frequent issues section, for people with questions
regarding the compilation of the project.

Removed libfreetype.a, as a solution was using the original FreeType
library. According to the FreeType docs, the glyph format we use cannot
be rendered using the normal function, so a workaround had to be
implemented. This workaround has been tested on Linux and Windows, but
still needs to be further tested on more systems.

ESRIMapDemo has been fixed. The version last pushed incorrectly
initialized pairs, and as a result did not compile on some systems.