add xp config, templated 1D stats
The project configuration now has listings for the src/xp folder,
allowing for developers to call add create object files and executables
from the xp folder. The folder also has its own makefile for managing
the temporary projects that exist. This makefile is not called from
other makefiles, but a build/xp folder does exist.

The one dimensional stats project is almost working. The median
algorithm has to be changed and the mode function needs to be fixed, but
everything else looks like it may be working. A test was written for
this and sample data is provided (mostly random).