Change AudioPlayer's contexts to a map
Change contexts to an unordered_map so that contexts can be accessed via
string that can act as names. The AudioPlayer destructor, newContext,
and setCurrentContext functions have been updated accordingly.

Add a command to all newly created contexts that ensures they don't try
to play video. Makes it so that links to places youtube-dl supports can
be played without issue.

Add a check to setVolume to make sure that the internal sprintf call can
never try to write to memory it shouldn't (the supplied volume should be
between 0 and 999).

Add functions to set a context to be playing or paused, but they are
unfinished, don't use them yet.