fix stock behavior, add debug configuration
launch.json:

launch.json now has a proper configuration for testing on Windows. Run
the 'g++ - run a.exe' task after running 'g++ -g <file>' to compile the
file to use VSCode's built-in debugger. This version currently requires
that the matching *.cpp file be open.

stock_delta.cpp:

A tldr (too long, didn't read) version of the following information is
present as comments in stock_delta.cpp

Stock no longer has a bit flag, to prevent underflow errors and save a
character per stock.

Minor bugs were fixed in the Stock constructor. These bugs should have
had no effect given that we overwrite the values most of the time, but
the fix was implemented just in case.

The first_day flag is now explicitly set every time new info is read
into a Stock object. This issue became incredibly apparent when the
tests we're running assumed that every Stock was a first stock, since
each stock shares a Stock object.

The delta function now accepts Stock pointers instead of Stocks, as
doing so allows us to default to nullptr.

The binwrite function was added to allow Compressed_stocks and Stocks to
be written directly to a binary file.