add CMake/Ninja to readme, add LZMA decompression
Adds a decompressor that uses the LZMA library from XZ Utils. Since
this uses XZ Utils, it actually takes an xz file with the default
configuration options. Do not try to give this test a .lzma file or
it will crash.

Adds lzmautil.hh, which holds functionality common between the various
lzma libraries that will be implemented in Grail.

Fixes readme to list information about installing and compiling with
CMake and Ninja.