Compile XDL code, add XDL tests
The XDL code now compiles. Adds a method of creating an XDLRequest
using existing XDLType pointers.

Makes Buffer's checkAvailableWrite method public. Need to check if a
method of getting that information already exists publicly.

Copies the Date and JulianDate types into std to convert them to XDL
types.

DynArray now converts the data pointer to a void or char pointer
(depending on the context) before performing any operations on it. This
is to allow manipulation of the array while preserving the constant
behavior of DynArrays of const objects.

XDL Lists have been abstracted out to List.hh. This is to prevent std.hh
from being flooded by list code.

Add Calendar type to the CompoundTypes.

Convert the readMeta used by SymbolTable.cc to the new readMeta.

Add a test for sending Stocks with the new List type.