Fix image being broken
The function setupBuffers() was created with default parameters, but for
some reason I was passing x, y, width, and height into that, which
caused the image I wanted to display to be tiled. The reason it appeared
to be a black square was because the image it was tiling was almost
entirely black (frame13.png), and when resized to as small as it was
none of the actual image comes through.

Close #32