Reparent StyledMultiShapes, redo elemPerVert
StyledMultiShape2D and StyledMultiShape25D are now subclasses of
MultiShape instead of MultiShape2D. This is mainly due to the transition
away using the Style class inside the MultiShapes.

A code review will be necessary to determine a) why MultiShape2D is
using styles still and b) if MultiShape2D is still necessary given that
StyledMultiShape2D is just MultiShape2D with RGB.

Also, each MultiShape (except for the base MultiShape class) implements
elemPerVert as a static constexpr, and a similar pattern should be used
for new MultiShape classes. This means that functions like
getPointIndex() can be called without any wrapper functions to account
for the differing numbers of elements per vertex point.