Massive overhaul of GraphWidget internals
This should have no impact on the actual user API, there's basically
only one change in the test files.

There's now a universal create axis function internal to GraphWidget
that performs differently based on what is allowed for a particular
graph and what is passed in. This is done with an enum with specific
values and using bitwise or to ensure valid input.

The internal function is never supposed to be called by the user, they
still call the create x and y axis functions which wrap the create axis
function.