changed NaN implementation used in GapMinderLoader
Some testing was done to see if Grail could be compiled using Clang
v12.0.1. While most of the project compiled just fine, there were some
minor issues where Clang saw the use of cmath's nan("") as something
that isn't a constexpr. Given that this was used in a constexpr
declaration (line 51 of GapMinderLoader.cc), this caused the rest of the
compile to be errored out. In C++ 11, the numeric limits library
introduced a templated quiet_NaN and signaling_NaN. These are both
constexpr by default and will insert NaN values (while giving the option
to check if there should be an exception on an NaN).

Warning: Anything using this data that is checking for NaN values should
either replace this with signaling_NaN or use std::isnan to check the
values of the data before doing any arthimetic operations.