Delete GLFW headers, add generic quantiles
The GLFW header files have been deleted from the include directory. GLFW
is a known system dependency and as of right now, including it is
just confusing GitHub into thinking Grail is a C project. There has
been a proposal to use a different GLFW implementation that is header
only, but that would likely be added as an external CMake dependency
rather than being included directly in the repository.

Also, the quantile functions used in Stats.hh have been replaced with a
hashtable to lookup any given quantile algorithm of the nine described
by Hyndman and Fan, 1996. This, as well as some setters and the ability
to set a default algorithm, allows users to pick the algorithm best for
their needs.