Revert to C++ 17, add fmtlib
Raspberry Pi does not seem to have a recent enough version of GCC to
support C++ 20. While this is not as much of an issue on Ubuntu (because
the Ubuntu toolchain repository exists), the hassle of using a newer
version of GCC on Raspberry Pi means its likely not worth using C++ 20
until there is better support.