Move contents of bin/.gitignore out to .gitignore
Previously if ./bin needed to be deleted, the user had to be careful not
to accidentally delete the .gitignore in ./bin as doing so would cause
any executables to be eligible to be added to the repo. Since ./bin
sometimes needs to be deleted in order to do a clean build (i.e. if
./build is deleted but the build mode needs to be changed), the contents
of ./bin/.gitignore have been moved to ./.gitignore so ./bin can be
completely deleted if necessary.