Update README.md based on #49
Also adds a .prettierrc.toml file, which will cap the length of lines in
Markdown files and wrap things to new lines to that it's much easier to
read, without changing what the final html/display on github looks like.
This assumes that you have the prettier utility installed in the first
place.