Update renderers, broke animation
Animation and updating are now split into different functions.
An update time (will be framerate) can be set so every dt seconds,
an update will be triggered. A render can be separately triggered,
but setDirty should trigger an update and render from now on.
Also, the stock server is rendering at least one page of data, but
more pages need to be implemented. Also a bug in the buffer reads needs
to be investigated, as large buffers with multiple sends do not get read
properly.

Animation needs to be fixed, as anything that previously animated will
most likely not animate.

Co-authored-by: Dov Kruger <hydrodog@users.noreply.github.com>