Callbacks and Animation are working again
Callbacks were being overwritten if they had an action number that was
too low, all actions now start at 1.

Globe and SolarSystem are now using the new grailmain system for
simplifying making programs with Grail.

There is a bug with Animation to do with using a frame rate that is too
high. An issue will be opened in order to resolve this later.

Co-authored-by: Dov Kruger <hydrodog@users.noreply.github.com>