Fix inconsistencies in SocketIO
SocketIO explicitly declared err_code as static, but this is already
implicity done due to err_code being in an anonymous namespace. The
explicit use of static has been removed.

SocketIO send and recv have new type signatures. Each function now
returns a ssize_t to be in line with Linux's function signatures. Given
that the functions originally returned an int. No functionality should
change. SocketIO recv used to take in a const char* buffer and cast to
a char*. Given that recv does modify the buffer, the buffer is now
explicitly a char* and then will be either kept a char* (on Windows) or
implicitly cast to a void* (on Linux). Given that void* and char* have
the same size, this should not present any issues.

SocketIO send and recv now use the flags provided. Previously, each
function passed 0 as the only flag. Be aware that Grail is designed to
be cross-platform and, as such, any flags used should be avaiable on
Windows or Linux.