Clean up minor issues
XDLRequest had an unnecessary import which was removed. The function
buildData is also commented out, as it is the base case for a function
which is also commented out.

GLWinFonts had a few loops where the counter is guarenteed to never be
negative, so int was replaced with uint32_t.

Callbacks had a static void function that was never given a void type
when it was declared. The implementation gave the correct type, but the
declaration needed to be fixed.