Change clang-tidy readability option
Currently, there are many places where an if-statement with only one
line following it is used without providing any braces. This is
primarily done out of convenience, however it also means many flags by
clang-tidy for readability issues. We solved this with indentation,
however if it ever becomes an issue we can always revert the setting so
any if-statement without braces triggers a warning.