Synchronizing older Windows commits
There are several modifications that were made for compatbility reasons,
however these changes are all several months old by now. These changes
are being added in all at once. Fixes will be made as the CI/CD system
reports errors.

AxisWidget: Include cstring to provide strlen

MultiText/MultiText2: Remove functions that have "unsigned long long" as
a parameter. Using functions with byte-specific types end up being
better for cross-platform compatbility.

Benchmark: Remove sys/times include. MSYS does not provide this and we
do not need it anymore.

DynArray: Add DynArray<T>::removeAt(uint32_t index) -> T. This removes a
particular element from the array and returns it. Warning: This
currently achieves this by copying the object to a temp variable,
sliding all other elements forward by one (without moving?), and then
returning the copy. For some reason, this invokes the original object's
destructor if its the last element in the array? I believe this needs
to be rewritten a bit before it is ready for use.

testDynArray: This runs through a couple test cases for DynArrays
(adding and removing integers, strings, and Tabs). This was originally
done as a way to test how adding and removing tabs worked out, but could
be useful in the future. This should eventually turn into a compile-time
test using static_assert, as DynArray should be able to be made into an
entirely constexpr class.

testfmt: Removes constexpr specifier on functions that call sprintf.
constexpr cannot be used on non-constexpr functions and sprintf is not
constexpr.