Remove libs folder, clean up CMakeFunctions
The libs folder should be generated during compile-time, and no artifacts should be kept. This has been reflected in the .gitignore.

GrailFunctions.cmake prints messages whenever one of the json files tracked is updated, but it previously added an extra newline.
Since this is inconsistent with the rest of CMake's styling, the newline has been removed.

Signed-off-by: Alice Huston <aliceghuston@gmail.com>