Refactor some XDL/IPV4Socket Code
This fixes some memory bugs in XDLRequest/IPV4Socket that cause
seg-faults on client side. There is still a bug in the StockQuoteDelta
tests where the client tries to reconnect before the server has closed
the connection, leading to a race condition. This will need to be solved
with multiplexing on the server-side.

Signed-off-by: ahuston-0 <aliceghuston@gmail.com>